Simple datatype adjustment.

+++ fertilized/leafs/classificationleafmanager.h
@@ -111,7 +111,7 @@ namespace fertilized {
       auto sample_list_ptr = data_provider.get_samples();
       for (const auto &sample_ptr : *sample_list_ptr) {
         if (! (sample_ptr.annotation[0] >= 0 &&
-               sample_ptr.annotation[0] < this -> n_classes)) {
+               static_cast<unsigned int>(sample_ptr.annotation[0]) < this -> n_classes)) {
             throw Fertilized_Exception("The annotations for classification "
               "must be in [0; n_classes)!");
         }
